+++ Sources/SwiftSDK/Persistence/DataQueryBuilder.swift
@@ -38,6 +38,7 @@ import Foundation
     public var related: [String]?
     public var groupBy: [String]?
     public var havingClause: String?
+    public var distinct = false
     
     var isRelationsDepthSet = false